Split fixture files for PSR-4 compliance
- Split Variables.php → Variables.php + ClosureVariables.php
- Split Enums.php → EnumUsage.php (uses existing Status/Priority enums)
- Split Inheritance.php → 8 individual class files
- Add Color enum to Fixtures\Enum for string-backed enum tests
- Update all test fixture paths
