+++ tests/Fixtures/src/Completion/ChildForSelf.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Fixtures\Completion;
+
+class ChildForSelf extends ParentForSelf
+{
+    public static string $ownProperty = 'child';
+
+    public static function ownMethod(): void
+    {
+        self::/*|self_inherited*/
+    }
+}
