+++ tests/Fixtures/src/Completion/ChildWithConstructor.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Fixtures\Completion;
+
+class ChildWithConstructor extends ParentWithConstructor
+{
+    public function __construct(string $name)
+    {
+        parent::/*|parent_access*/
+    }
+}
