+++ tests/Fixtures/src/Completion/ChildWithPrefix.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Fixtures\Completion;
+
+class ChildWithPrefix extends ParentWithMethods
+{
+    public function test(): void
+    {
+        parent::gr/*|parent_prefix*/
+    }
+}
