+++ tests/Fixtures/src/Completion/ClosureVariables.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Fixtures\Completion;
+
+class ClosureVariables
+{
+    public function withClosure(): void
+    {
+        $fn = function ($param) {
+            $localVar = 1;
+            $l/*|closure_local*/
+        };
+    }
+}
