+++ tests/Fixtures/src/Completion/NoParent.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Fixtures\Completion;
+
+class NoParent
+{
+    public function test(): void
+    {
+        parent::/*|parent_no_parent*/
+    }
+}
