+++ tests/Fixtures/src/Completion/ParentWithConstructor.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Fixtures\Completion;
+
+class ParentWithConstructor
+{
+    public function __construct(string $name)
+    {
+    }
+
+    protected function greet(): string
+    {
+        return 'Hello';
+    }
+}
